requirements.txt
openapi
requests
python-dotenv
pydantic
